requirements.txt
openai==1.107.3
python-dotenv==1.0.1